Refactor UploadFile component
