+++ src/util/util.js
@@ -10,3 +10,14 @@ export const filterRepositoriesNames = (repositories) => {
 
   return filtredRepositories.map((repository) => repository.full_name)
 }
+
+export const formatFileName = (fileName) => {
+  const charLimit = 30
+  const stringList = fileName.split('.')
+
+  if (stringList[0].length > charLimit) {
+    return stringList[0].slice(0, charLimit) + '...' + stringList.at(-1)
+  }
+
+  return fileName
+}
